services/ai-suite/all_legacy_code/requirements/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=3.5.0=py311h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py311h5eee18b_0
array-record=0.5.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=py311h06a4308_0
attrs=23.1.0=py311h06a4308_0
babel=2.11.0=py311h06a4308_0
beautifulsoup4=4.12.2=py311h06a4308_0
bleach=4.1.0=pyhd3eb1b0_0
brotli-python=1.0.9=py311h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=py311h06a4308_0
cffi=1.16.0=py311h5eee18b_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.1.2=py311h06a4308_0
contourpy=1.2.0=pypi_0
cryptography=41.0.7=py311hdda0065_0
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.28=h52b45da_1
datasets=2.16.1=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py311h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.7=pypi_0
distro=1.9.0=pypi_0
dlib=19.24.2=pypi_0
dm-tree=0.1.8=pypi_0
etils=1.6.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.5.0=h6a678d5_0
face-recognition=1.3.0=pypi_0
face-recognition-models=0.3.0=pypi_0
filelock=3.13.1=pypi_0
flatbuffers=23.5.26=pypi_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.47.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2023.10.0=pypi_0
gast=0.5.4=pypi_0
glib=2.69.1=he621ea3_2
google-api-core=2.15.0=pypi_0
google-auth=2.26.1=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
google-cloud-texttospeech=2.15.1=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.62.0=pypi_0
grpcio=1.60.0=pypi_0
grpcio-status=1.60.0=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.20.2=pypi_0
icu=73.1=h6a678d5_0
idna=3.6=pypi_0
importlib-resources=6.1.1=pypi_0
ipykernel=6.25.0=py311h92b7b1e_0
ipython=8.20.0=py311h06a4308_0
ipywidgets=8.0.4=py311h06a4308_0
jedi=0.18.1=py311h06a4308_1
jinja2=3.1.2=py311h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py311h06a4308_0
jsonschema-specifications=2023.7.1=py311h06a4308_0
jupyter=1.0.0=py311h06a4308_8
jupyter-lsp=2.2.0=py311h06a4308_0
jupyter_client=8.6.0=py311h06a4308_0
jupyter_console=6.6.3=py311h06a4308_0
jupyter_core=5.5.0=py311h06a4308_0
jupyter_events=0.8.0=py311h06a4308_0
jupyter_server=2.10.0=py311h06a4308_0
jupyter_server_terminals=0.4.4=py311h06a4308_1
jupyterlab=4.0.8=py311h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.25.1=py311h06a4308_0
jupyterlab_widgets=3.0.9=py311h06a4308_0
keras=2.15.0=pypi_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h143b758_1
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.6=pypi_0
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm14=14.0.6=hdb19cb5_3
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.1=pypi_0
markupsafe=2.1.3=py311h5eee18b_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=py311h06a4308_0
mistune=2.0.4=py311h06a4308_0
ml-dtypes=0.2.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
mysql=5.7.24=h721c034_2
mysql-connector-python=8.2.0=pypi_0
nbclient=0.8.0=py311h06a4308_0
nbconvert=7.10.0=py311h06a4308_0
nbformat=5.9.2=py311h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py311h06a4308_0
nltk=3.8.1=pypi_0
notebook=7.0.6=py311h06a4308_0
notebook-shim=0.2.3=py311h06a4308_0
numpy=1.26.3=pypi_0
nvidia-cublas-cu12=12.2.5.6=pypi_0
nvidia-cuda-cupti-cu12=12.2.142=pypi_0
nvidia-cuda-nvcc-cu12=12.2.140=pypi_0
nvidia-cuda-nvrtc-cu12=12.2.140=pypi_0
nvidia-cuda-runtime-cu12=12.2.140=pypi_0
nvidia-cudnn-cu12=8.9.4.25=pypi_0
nvidia-cufft-cu12=11.0.8.103=pypi_0
nvidia-curand-cu12=10.3.3.141=pypi_0
nvidia-cusolver-cu12=11.5.2.141=pypi_0
nvidia-cusparse-cu12=12.1.2.141=pypi_0
nvidia-nccl-cu12=2.16.5=pypi_0
nvidia-nvjitlink-cu12=12.2.140=pypi_0
oauthlib=3.2.2=pypi_0
openai=1.7.1=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.12=h7f8727e_0
opt-einsum=3.3.0=pypi_0
overrides=7.4.0=py311h06a4308_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.2.0=pypi_0
pip=23.3.2=pypi_0
platformdirs=3.10.0=py311h06a4308_0
ply=3.11=py311h06a4308_0
prometheus_client=0.14.1=py311h06a4308_0
promise=2.3=pypi_0
prompt-toolkit=3.0.43=py311h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
proto-plus=1.23.0=pypi_0
protobuf=4.21.12=pypi_0
psutil=5.9.0=py311h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=14.0.2=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyaudio=0.2.14=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pydub=0.25.1=pypi_0
pygame=2.5.2=pypi_0
pygments=2.15.1=py311h06a4308_1
pyopenssl=23.2.0=py311h06a4308_0
pyparsing=3.1.1=pypi_0
pyqt=5.15.10=py311h6a678d5_0
pyqt5-sip=12.13.0=py311h5eee18b_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.7=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py311h06a4308_0
python-json-logger=2.0.7=py311h06a4308_0
pytz=2023.3.post1=py311h06a4308_0
pyyaml=6.0.1=py311h5eee18b_0
pyzmq=25.1.0=py311h6a678d5_0
qt-main=5.15.2=h53bd1ea_10
qtconsole=5.5.0=py311h06a4308_0
qtpy=2.4.1=py311h06a4308_0
readline=8.2=h5eee18b_0
referencing=0.30.2=py311h06a4308_0
regex=2023.12.25=pypi_0
requests=2.31.0=py311h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=py311h06a4308_0
rfc3986-validator=0.1.1=py311h06a4308_0
rpds-py=0.10.6=py311hb02cf49_0
rsa=4.9=pypi_0
safetensors=0.4.1=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
send2trash=1.8.2=py311h06a4308_0
setuptools=69.0.3=pypi_0
sip=6.7.12=py311h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py311h06a4308_1
sounddevice=0.4.6=pypi_0
soupsieve=2.5=py311h06a4308_0
speechrecognition=3.10.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.15.0.post1=pypi_0
tensorflow-datasets=4.9.4=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.35.0=pypi_0
tensorflow-metadata=1.14.0=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.17.1=py311h06a4308_0
threadpoolctl=3.2.0=pypi_0
tinycss2=1.2.1=py311h06a4308_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.0=pypi_0
toml=0.10.2=pypi_0
tornado=6.3.3=py311h5eee18b_0
tqdm=4.66.1=pypi_0
traitlets=5.7.1=py311h06a4308_0
transformers=4.36.2=pypi_0
typing-extensions=4.9.0=pypi_0
typing_extensions=4.7.1=py311h06a4308_0
tzdata=2023.4=pypi_0
urllib3=2.1.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py311h06a4308_1
websocket-client=0.58.0=py311h06a4308_4
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py311h06a4308_0
widgetsnbextension=4.0.5=py311h06a4308_0
wrapt=1.14.1=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.5=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
